Make generated code for resources safer.
Give the option for the generated code never to return null strings, but
return string.Empty instead:

    ConvertResxStringsToCsharp `
        -Path Strings.resx `
        -Destination Strings.cs `
        -Namespace MyNamespace `
        -ClassName MyClass `
        -NoNullStrings

Make sure that the generated code always has a section for invariant
culture. If the user does not specify a resx for invariant culture,

    ConvertResxStringsToCsharp `
        -Path @{'en-US'='us.resx'; 'en'='en.resx'; 'de-BE'='be.resx'; 'ja'='ja.resx';} `
        -Destination Strings.cs `
        -Namespace MyNamespace `
        -ClassName MyClass `
        -NoNullStrings

then choose one for invariant culture, using the culture that is first
if the culture names were sorted by length and by name. In the example
above, 'en' would also be used as invariant culture because out of the
shortest culture names above ('en' and 'ja'), 'en' comes first in
alphabetical order.