Optimize string resources.
Make the arrays of localized strings lazily.

Since it is very likely that at most only 2 or 3 cultures will be used
during the lifetime of the program (such as en-US, en, invariant, or
de-DE, de, invariant, and so on) it would be wasteful to take the time
to create all the arrays of localized strings all at once.