Optimize string resources.
For example:

    //Strings class is generated
    Strings.Culture = new CultureInfo("en-US");
    Console.WriteLine(Strings.SomeLocalizedString);
    Console.WriteLine(Strings.AnotherLocalizedString);
    Console.WriteLine(Strings.YetAnotherLocalizedString);

Each localized string read causes the Strings class to perform calls to
the Dictionary.ContainsKey method 1 or more times (for en-US culture and
maybe en culture).

Move the calls to Dictionary.ContainsKey method when the Culture is set.
This is a good trade-off because the Culture is typically only set once,
while localized strings are read many times during the lifetime of the
program.