Optimize string resources.
Change foreach loop on List<Lazy<string[]>> to a for loop on
Lazy<string[]>[].

A foreach loop on a List<T> has a try/finally and dispose. A for loop on
a List<T> has a getItem function call (for indexing into List<T>). I'm
not sure what is the difference between a for and foreach loop on an
array, but, let's just go with the most basic one.