-nix-ification
Make sure comparisons with path strings are using the proper directory
separator characters.

In Windows, the directory separator characters can be "\" or "/", but
mostly uses "\". Since powershell script writers are writing paths using
"/" in Windows in order to be compatible with Linux, we have to make
sure that the paths we are comparing are using the same directory
separator characters.